Added 'path' to EWD.setCookie function
When a user of an SPA hits page refresh or uses a bookmark as an entry
point into the app, it is necessary to call the EWD.setCookie function
in order to re-register the session and to map the session cookie to
the new web socket.

If the href has a different path to that when the token was first
generated, you can end up in a situation where QEWD loses track of the
session because it is seen as a new token.

This modification allows one to prevent this undesirable behaviour by
providing the ability to specify the cookie’s path and providing a
sensible default for an SPA